Fully implemented Green Screening on Server
Now that the colour spaces are working together correctly, I have fully implemented green screening on the server. I have included a HD and SD version of the green screening function (SD take 5 seconds to process, HD takes 8 seconds to process). Currently, this is all still based off the image existing on the server, but will be adaptable to images being sent to the server aswell.