Successfully sending user to authorise Instagram login
Currently, as the login is being done in the main activity of the app, on returning after authorisation the user is being prompted to login again. Will split into two activities. It is only now that I really see what the restriction of Instagram's 'Sandbox' is going to have on our app. When I tried to authorise the app using my personal instagram account, I got an error in the browser saying 'you are not a sandbox user of this client'. This means we would have to individually invite any users that we want to allow to access Instagram through our app, up to a maximum of 10 users. This limit also effects the number of apps you can develop, as each user can only be a sandbox user on a maximum of 5 apps. In order to move past this sandbox restriction, we would need to submit our app to Instagram for approval.