Improvement on y axis
Had a mad idea last night about the movement. The movement is minimal on the x axis and the y axis but if i multiple the values it will have a greater effect I found this to be true and most effective for the y axis but unfortunately it is not doing anything for the x axis which does move but its very minimal. Also there seems to be and error that causes the prototype to crash, it is a fatal exception to the animation thread. The error stated is an illegal Argument Exception and it says - File https://maps.googleapis.com/maps/api/streetview?location=40.7585806,-73.9850687&heading=-0.42255324&pitch=-2.5609288&key=&size=480x800 contains a path separator