Better y, still trying to fix a-xais
The movement of the y axis is better but needs to be inverted. Originally it had displayWidth as the second stop point which when looking back at it wouldnt work and could be one of the things preventing me from seeing beyond the width of the deivce when the sensor detects an accelerometer event. So i got rid of that and it seems to make the x axis flow better but Im still not able to actually see 360 of the images, probably more so a 90 degree segment. As for the Y it needs to be inverted as the ground is shown if the device is tilted up and it shows up if the device is titled down. I've tried making the mapping figures oppisite and the accelerometer event for accelerometerY I tried making it a minus number by multipling by -2 and by applying extra brackets to see would it enforce the minus number but currently has no effect. Even when I print of the figures I can see that yes they do need to be inverted but are not being inverted. For example, when the device is up the number printed is close to ten while down it prints out to 1. I need to invert these to make sure the action is normal for the user