Fixed accelerometer
I noticed with the accelerometer that it had a tendancy to revert back to a point. It no longer does this and now I have tried to implement the accelerometer and the mouse press for moving in the environment, so far the accelerometer works fine but now the mouse press reverts to a point. My thinking is that as the mouse press is concerned with the previous mouse x and mouse y and getting the distance between the lastest mouse x and mouse y, this is conflicting with the accelerometer as the accelerometer has moved past where the mouse x and mouse y are but it is still focused on the mouse x and mouse y, I tried tying it to the accelerometer x but it doesnt have any effect. I will have to look at how the previous mouse x and previous mouse y are and try update their values or something because they are incorrect with the values of the accelerometer