Fix to saving favourites and removal of camera live view title
Fixing saving of favourites, as when a favourite was removed from the favourites xml, was not resaving the xml, so the next time it checked was it a favourite it was not picking it up. Saving of favourites now fully functional. Also removing title from camera live view as we have decided that it is not necessary