Implemented Saving locally, autosave mode on/off
Implemented saving locally (for sharing images to Twitter). Created
seperate functions for saving to the user's photo gallery or saving to
the app's local storage. Allowing user to turn on/off autosave mode.
Using booleans to determine the appropriate message to display on the
successful page i.e. saved, shared or saved and shared.