Established which method for orientation
As we want to detect which orientation the device is currently at, I decided to test some of the other different methods that we could possibly use for this dectection. There are methods such as gyroscopevents and onGameRotationEvents but these are things that need to be built into a device and we dont know what device the usre will using so we cant limit ourselves to it. Also the device im using for testing doesnt support them so its a good indication showing that not all the user will have these features. This sketch was testing the different methods but the accelerometer seems to be the best and provides results as some of the othr methods were'nt returning anything or printing out in the console