All response content being read as string
I just read through the code for the 'HTTP Requests for Processing'. All responses from the server are being parsed as String, so no matter what I do server side, the result will always be a string. Based on this, if I want the image to be send in the response, I am going to have to write my own HTTP request, without this library.