Optimise - Reduced memory load, increased cpu load
In an attempt to reduce the memory load, I am now storing the paths to the image, instead of the images themselves. Although this has reduced the memory load from 44.68MB to 19.56MB, the CPU usage has now gone through the roof, so the app is almost unresponsive (as each image is having to be loaded in repeatedly).