Loading icons in as PImages
As a comprimise, to try and balance the load between the memory and the CPU, I am loading only the icons back in as PImages (the rest of the images will be loaded in as needed). Haven't implemented using these images yet, so the app still runs slow, but I just wanted to check how the memory would be effected by these PImage variables. It has increased by 10MB, and so is currently sitting around 30MB (which appears to be 2MB below the allocated memory), but the original app was running at 44MB, so there is still a 14MB improvement.