Fix to save image before sending to server
The ketai .savePhoto method I was using was not consistent, sometimes it worked, sometimes it didn't. Now storing the ketaiCamera image in a PImage variable (using the .get() method to get the image from the ketaiCamera object). Then saving this image using the processing .save() method. Keying now works fully on the server, but takes about 20seconds per image