Added Sharing options to app
Adding the option to share the image outside of the app. This button appears after saving to twitter and/or saving to the device takes place (as once the user leaves the app, the app will reset, so these tasks need to be completed first). Repositioning icons on screens to accomodate for this. If a user doesn't save or share their image, they are taken back to the camera live view screen. Since the image has to exist in an external folder in order for it to be shared by an external app, saving the image (if it was not already saved) before directing them to the device's sharing menu. This sharing menu contains each of the apps which the user can share the image to (varies depending on the apps installed on the device).