Huge improvement in speed/memory/overall running of app
We decided to remove the backgroun image from all screens, and instead just display one common screen background (the general page background) to reduce the load on the server. Then removed all references to colors from constructors within the app, as the majority of elements in the app no longer require a color (appart from text inputs and favourite tabs). Using a setBackgroundColor() to set these, should they be needed. We then endcountered a major error, which caused the app to flash extensively and crash. The following error message was logged out to the console repeatedly before the crash - 'W/art: Failed to open zip archive '/system/framework/WfdCommon.jar': I/O Error'. This appeared to be a processing error. This error persisted for over an hour and a half, as we tried to retrace our steps to see what could have gone wrong. I then realised that I had removed the .get() methods from the ketaiCamera and currentFrame instances, to try and reduce the load on memory (as this loads the pixels array of the camera, from what I can see in the logcat). Once I reverted this to the original method, the error slowly began to resolve itself. First it stopped flashing, but kept logging the error. Then it began to run more efficently than it had in a long time (i.e. CPU and memory speeds were good) and finally, the application ran on the Google Nexus tablets, which it has not been able to do in almost 8 weeks (due to 'our of memory' errors). The app is now running pretty well on all the devices we have tested, with the green screen keying being the only slow part of the interaction. BUGS (Images not saving to Nexus. Live image from camera not refreshing when you return to camera live view screen after sending/sharing an image.