Optimising - loading images into construcotrs
Optimisng the app by loading all icon image directly into their constructors (so that these images do not end up being duplicated in the main app. Moving the declaration of the currentFrame variable into the green screen keying function, so that they are not declared globally. Setting these to null once the thread has completed. Fixed positioning of the overlay image on the image the user takes. Removed the images from buttons, and replaced them with just rectangles, with white fills and a black stroke.