Removing currentFrame var - breaking green screen
We tried removing the currentFrame variable, to try and reduce the amount of variables in the app. This is completly breaking the green screen - not always right away, but after taking a picture and returing to it, stripes appear all over the image, and the colours get distorted. Saving this as a prototype, for reference