Additional graphics (and functionalities) added
Added the ability to turn on/off the saving of an individual image
(regarless of what the autosave setting is). Added the ability to turn
on/off the sharing of an individual image to Twitter (only possible if
the user has already logged in with their twitter account - otherwise it
defaults to off and is not changeable).
If both twitter and saving are unticked in SaveShareScreenA, then
changing the "next" button to a "done" button, so that the user is
simply taken back to the camera live view, as opposed to continuing
through the save/share route. When the app first loads, setting the
default images for the save and twitter share icons to corrospond with
the current settings (i.e. whether autosave mode has been enabled in the
user's preferences, which are loaded in from an external XML file, or
whether they have logged in with Twitter).
Setup the basic on/off functionality for the learning mode, so that it
can also be stored in the user's preferences. Defaulting the images for
the toggle switches on the settings screen to illustrate their
respective setting's current status when the sketch loads.
Made all icons that need to have their images swapped out at some point
public, so that their .setImage() methods (inherited from the Rectangle
class) can be called on them directly from the main sketch.
Made the text on the ShareSaveUnsuccessfulScreen dynamic, so that it
will change to reflect which of these (if not both) caused an issue, and
offer the relevant options to the user i.e. to retry etc.
Changed the setup of the sketch, so that the connection to the Twitter
API only occurs if the user is logged in to Twitter, to try and reduce
the overhead of loading the app.
Setting the loading screen to null as soon as it is no longer needed, as
it can only ever be seen once per app session anyway.
Adding in the on/off images for shakeMovement on/off, so that users can
tell which mode they are in.
Including an additional graphic for the switchCameraView icon, so that
if this app is run on a device that only has one camera, this option
will appear as having been disabled (as switching between cameras will
not be possible).
Added a try/catch to the mergeImages function, as this seems to be the
most common place for out of memory errors to occur, now that we have
resolved this issue in the camera live view screen. If an error occurs,
redirecting the user to the relevant share save unsuccessful screen,
where they will be given the relevant options (i.e. try again, share
etc).
Adding the current user's account name to the social media logout
screen, so they can see which account is logged in. The storing/removing
of Twitter accounts has not yet been implemented.
Re-exported many of the larger graphics in the app, as they were much
larger than they needed to be. The app is now running on approximately
33MB memory (down from 45MB earlier this week).
There seems to have been an ongoing issue with the Twitter login button
(on the first screen) for a while, but it was only today we realised the
significance of it. The button sometimes would appear grey, and now does
not function at all. The app was still detecting clicks (logging them as
ViewPostImeInputStage ACTION_DOWN) but is not triggering the button. I
was still able to trigger the login event by calling
loginButton.performClick() to imitate a click event. Re-wrote the
twitter activity's layout .xml file, storing the text and alignment
values in the relevant dimens and string res files. The button is back
working now, but I am still at a loss to what the issue was, as I don't
think I made any significant changes when fixing it.