Re-commenting code (+ minor fixes)
No longer storing the keyed image from ketai camera as a
"rectBackroungImage" in the Rectangle class, as this also required it's
scaleX and rotation to be stored in two more variables, which meant that
the Rectangle class, which every other class inherits from, had three
variables which were completely specific to one image on one screen.
Instead just adding the keyed image using the rectangle class's
addImage() method, just like every other image in the app.
Re-commenting the Rectangle class fully, as well as relevant portions of
other sub classes