More re-commenting of main app
Changed to using ternary operators to decide the default image for icons
on their individual screen's pages (cleaner). Removed all variables,
constructor parameters and methods relating to password from the
TextInput class, as we are no longer dealing with users passwords
directly in the app (this is all handled either through OAuth redirects
for Twitter, of through the individual apps themselves when using share
intents).