Continuing commenting of main code (+ scrolling fix)
Continuing commenting of main code. Fixed the way in which screens are
scrolled, so that now this is based on a method inherited from the
Screen class, as opposed to having the same code duplicated in screens
that need to be scrollable. Now, if a screen needs to be scrolled, can
just call the scrollScreen() method on it.