Continuing commenting (+random location xml)
Continuing with the recommenting of the code in the main app. Also
creating a new XML file to store the random locations for the app.
Previously, these were stored in the Favourites screen as a String[]
array, and then when we started using XML for storing user preferences
they were being accessed from this file, but now favourite locations
have their own XML file, so that even if a user removes all the default
locations from their favourites, they will still be available for
accessing random locations within this file.