Successfully intergrated learning mode
The learning mode has been successfully implemented into the main android studio file. Small comments were adjusted/made. The is now a textbox method which makes it easier to make text boxes. Also on scrolling screens the ability to tap and be brought back to the top of the page has been implements