Orientation of google images and preview images
The orientation of the google street view images can be detected, so when the device is turn the google stret view images are turned also. Once a picture is taken and compiled it too will have the correct orientation of the google street view images in the camera preview screen. Also the image itself will be the right orientation. The overlay also has been rotated if he image is landscape