Finalizing and fixing image save rotation
Fixed how the image is displayed so the google image rotates based on orientation, this works also for the other camera on the device. Also the overlay is approciately positioned for both landscape and portrait images and appears for the user. The user can save out portrait and lascape images and they will be sized appropriately