Compiling-Sprint-012-Production-20-ReRelease
In light of our first offical release, we were met with a few things that needed to be fixed. Now the app has the ability to prompt users to use their camera and storing files to their device, this is important for any phone that is android version 6.0 or marshmallow. Also this version has the ability to scroll around on the live camera view in landscape mode