Storing Twitter Login details
Creating a new TwitterLoginActivity_CheckLogin class, which contains a
separate Processing sketch to the rest of the application. Using the
class to check if a user is already logged in to Twitter, by checking
their user_preferences.xml file. If the user is already logged in, then
passing these values back to the relevant TwitterLoginActivity static
variables, so that they can be accessed from within the main sketch, and
calling the activity's goToMainActivity() method, so the user does not
have to go through the twitter login screen again, as they are already
logged in.
When a user logs in to Twitter with a new account, storing these details
in the user_preferences.xml file, so that next time they open the app,
these details will be persisted.
NOTE - not storing a user's password, just storing the access tokens,
userId and username. Details are never stored remotely, and exist only
in the user_preferences.xml file which is stored locally to the app, so
cannot be accessed from other applications on the device either.