Minor fixes
If a user presses the back button on their device, it was being detected
as a key press, and was causing an error is a current text input didn't
exist. Now checking if an input exists before any key is detected. Also
adding an onStop() override, to ensure the ketai camera is stopped when
a user leaves the app. Still can cause a crash on resume due to out of
memory errors.