Removed the learningScreen class
This is a prototype removing the learningscreen class and making the green screen toast alerts only appear when learning mode is on. This decision has been implemented as the old learning mode appeared too glichy and was flashing, to combat this we have chnaged how the learning mode happens and what it is for ie. alerting the user they need green for the app not for the apps functionality