Requesting camera permissions, but only front camera is required
Previous commit (where no camera permissions were requested, and rear camera was not required) resulted in the app being compatible with the Google Nexus 7 tablets for the first time since we started working on this project. When the app downloaded however, the camera was not working (as no permissions were being requested). Instead, we are now requesting the permissions for the camera, but setting the use-feature for the back camera to false, and the front to true (so devices need to have at least one camera to use our app).