Tidy Up
Removing the 'TeamProject_Submission' folder, as it only contains some of the components we actually submitted for the project (the actual components are too large to be synced up to GitHub). Also placing the 'Wish_I_Was_Here' folder in a new directory, as this is just the original app in Processing, while the final project exists in the 'AndroidStudio' folder.